pyproject.toml
[project]
name = "gps-test"
version = "0.1.0"
requires-python = "==3.10.*"
dependencies = [
    "pyserial>=3.5",
    "pynmea2>=1.19",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
]
